+++ src/utils.js
@@ -1,7 +1,4 @@
-const buffer = require('most-buffer');
 const luxon = require('luxon');
-const most = require('most');
-const nth = require('most-nth');
 
 
 function formatTimezone(offset) {
@@ -11,6 +8,8 @@ function formatTimezone(offset) {
   return (Math.sign(offset) ? '+' : '-')
     + [hours, offsetValue - hours * 60].map((value) => String(value).padStart(2, 0)).join(':');
 }
+
+function isNotUndefined(value) { return value !== undefined; }
 
 function isNull(value) { return value === null; }
 
@@ -24,14 +23,7 @@ function isPredictiveModel(value) {
     && value.configuration !== null;
 }
 
-function isStream(value) {
-  return value !== null
-    && typeof value === 'object'
-    && typeof value.pipe === 'function'
-    && value.readable !== false
-    && typeof value._read === 'function'
-    && typeof value._readableState === 'object';
-}
+function notString(value) { return typeof value !== 'string'; }
 
 function parseDate(value) {
   if (value === null || value === undefined || typeof value === 'boolean') return;
@@ -53,48 +45,16 @@ function parseTimestamp(value) {
   throw new Error();
 }
 
-function toBuffer(stream) {
-  return stream
-    .thru(buffer())
-    .thru(nth.first)
-    .then((buffer) => buffer || []);
-}
-
-function toStream(value) {
-  if (isStream(value)) return fromStream(value);
-
-  try {
-    return most.from(value);
-  } catch (error) {
-    return most.throwError(error);
-  }
-}
-
-
-function fromStream(value) {
-  return most
-    .fromEvent('data', value)
-    .merge(most
-      .fromEvent('error', value)
-      .chain(most.throwError))
-    .until(most
-      .fromEvent('end', value)
-      .take(1)
-      // Prevent `end` event to close the stream before `error` event
-      .delay(100));
-}
-
 
 const DateTime = luxon.DateTime;
 
 
 module.exports = {
   formatTimezone,
+  isNotUndefined,
   isNull,
   isPredictiveModel,
-  isStream,
+  notString,
   parseDate,
   parseTimestamp,
-  toBuffer,
-  toStream,
 };