+++ test/helpers/index.js
@@ -73,7 +73,6 @@ function registerEndpoint() {
 }
 
 
-const INPUT_METHODS = [identity, streamify];
 const INVALID_DATES = [false, NaN, 'N/A', 'NaN', 'unknown', '123456', 'string', '5151-51-51T51:51:51.515Z'];
 const INVALID_NUMBERS = [null, {}, '12', true, new Date];
 const INVALID_OBJECTS = [null, 0, true, 'string', Symbol()];
@@ -86,7 +85,6 @@ module.exports = {
   destroyContext,
   identity,
   streamify,
-  INPUT_METHODS,
   INVALID_DATES,
   INVALID_NUMBERS,
   INVALID_OBJECTS,